### 2021-05-21 #3301 @whedon
[PRE REVIEW]: LDDS: Python package for computing and visualizing Lagrangian Descriptors for Dynamical Systems
**Submitting author:** @Shibabrat (<a href="http://orcid.org/0000-0001-7964-2513">Shibabrat Naik</a>)
**Repository:** <a href="https://github.com/champsproject/ldds" target ="_blank">https://github.com/champsproject/ldds</a>
**Version:** v0.1.0
