### 2021-05-24 5c4bbdf
Added JOSS badge and centred image captions
