+++ paper/paper.bib
@@ -33,6 +33,7 @@
 	number={1-2},
 	pages={55--81},
 	year={1984},
+    doi = {10.1016/0167-2789(84)90270-7},
 	publisher={Elsevier}
 }
 
@@ -55,7 +56,7 @@
 	volume = {149},
 	pages = {413-428},
 	year = {2018},
-	doi = {https://doi.org/10.1016/j.ijmecsci.2017.10.040},
+	doi = {10.1016/j.ijmecsci.2017.10.040},
 	author = {Zhong, J. and Virgin, L. N. and Ross, S. D.}
 }
 
@@ -65,7 +66,7 @@
 	volume = {82},
 	pages = {105033},
 	year = {2020},
-	doi = {https://doi.org/10.1016/j.cnsns.2019.105033},
+	doi = {10.1016/j.cnsns.2019.105033},
 	author = {Zhong, J. and Ross, S. D.}
 }
 
@@ -91,11 +92,12 @@
 }
 
 @book{wiggins1994normally,
-	title={Normally hyperbolic invariant manifolds in dynamical systems},
-	author={Wiggins, S.},
-	volume={105},
-	year={1994},
-	publisher={Springer Science \& Business Media}
+	title= {Normally hyperbolic invariant manifolds in dynamical systems},
+	author= {Wiggins, S.},
+	volume= {105},
+	year= {1994},
+	doi = {10.1007/978-1-4612-4312-0},
+	publisher= {Springer Science \& Business Media}
 }
 
 @article{OzoriodeAlmeida1990,
@@ -104,6 +106,7 @@
 	title = "Geometry and dynamics of stable and unstable cylinders in 
 	Hamiltonian systems",
 	journal = "Physica D: Nonlinear Phenomena",
+    doi = "10.1016/0167-2789(90)90040-V",
 	volume = "46",
 	number = "2",
 	pages = "265 - 285",
@@ -123,6 +126,7 @@
 	title={An analytical study of transport, mixing and chaos in an unsteady vortical flow},
 	author={Rom-Kedar, V and Leonard, A and Wiggins, S},
 	journal={Journal of Fluid Mechanics},
+    doi={10.1017/S0022112090000167},
 	volume={214},
 	pages={347--394},
 	year={1990},
@@ -232,6 +236,7 @@
 	volume = {18},
 	number = {12},
 	journal = {Communications in Nonlinear Science and Numerical Simulation},
+    doi = {10.1016/j.cnsns.2013.05.002},
 	author = {Mancho, A. M. and Wiggins, S. and Curbelo, J. and Mendoza, C.},
 	month = dec,
 	year = {2013},
@@ -240,14 +245,14 @@
 
 @article{katsanikas2020a,
 	Author = {Katsanikas, M. and Garc\'{i}a-Garrido, V. J. and Wiggins, S.},
-	Doi = {https://doi.org/10.1016/j.cplett.2020.137199},
+	Doi = {10.1016/j.cplett.2020.137199},
 	Issn = {0009-2614},
 	Journal = {Chemical Physics Letters},
 	Pages = {137199},
 	Title = {The dynamical matching mechanism in phase space for caldera-type potential energy surfaces},
 	Volume = {743},
-	Year = {2020},
-	Bdsk-Url-1 = {https://doi.org/10.1016/j.cplett.2020.137199}}
+	Year = {2020}
+    }
 
 @article{GG2020a,
 	Author = {Garc\'{i}a-Garrido, V. J. and Naik, S. and Wiggins, S.},
@@ -262,14 +267,14 @@
 
 @article{GG2020b,
 	Author = {Garc\'{i}a-Garrido, V. J. and Agaoglou, M. and Wiggins, S.},
[…]
