### 2021-05-28 #3301 @broncio123
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1103/RevModPhys.64.795 is OK
> - 10.1111/j.1365-2966.2005.09237.x is OK
> - 10.1063/1.166509 is OK
> - 10.1063/1.4915831 is OK
> - 10.5281/zenodo.4539666 is OK
> - 10.1007/BF00375090 is OK
> - 10.1134/S1560354716060034 is OK
> - 10.5281/zenodo.3958985 is OK
> - 10.1016/j.compfluid.2016.07.007 is OK
> - 10.1016/j.jocs.2014.12.002 is OK
> - 10.1063/1.4795749 is OK
> - 10.1017/jfm.2020.737 is OK
> - 10.1142/S0218127420300086 is OK
> - 10.1142/S0218127420500765 is OK
> - 10.1088/1751-8121/ab8b75 is OK
> - 10.1039/D0CP01362E is OK
> - 10.1007/s00382-019-04832-y is OK
> - 10.1007/s00382-019-04833-x is OK
> - 10.1103/PhysRevE.100.022204 is OK
> - 10.1142/S021812741750225X is OK
> - 10.1016/j.cnsns.2019.104907 is OK
> - 10.1175/JAS-D-11-0142.1 is OK
> - 10.1134/S1560354718050052 is OK
> - 10.1016/j.marpolbul.2016.08.018 is OK
> - 10.4249/scholarpedia.6327 is OK
> - 10.1002/9780470977859 is OK
> - 10.1016/j.physd.2005.10.007 is OK
> - 10.1016/j.cnsns.2015.02.022 is OK
> - 10.1142/S0218127417300014 is OK
> - 10.1038/s41598-018-23028-8 is OK
> - 10.1088/1751-8121/ab8b75 is OK
>
> MISSING DOIs
>
> - 10.1201/9781003069515-39 may be a valid DOI for title: Transport in Hamiltonian systems
> - 10.1007/978-1-4612-4312-0 may be a valid DOI for title: Normally hyperbolic invariant manifolds in dynamical systems
> - 10.1016/0167-2789(90)90040-v may be a valid DOI for title: Geometry and dynamics of stable and unstable cylinders in  Hamiltonian systems
> - 10.1017/s0022112090000167 may be a valid DOI for title: An analytical study of transport, mixing and chaos in an unsteady vortical flow
> - 10.1016/j.cnsns.2013.05.002 may be a valid DOI for title: Lagrangian Descriptors: A method for Revealing Phase Space Structures of General Time Dependent Dynamical Systems
> - 10.1175/jas-d-12-0274.1 may be a valid DOI for title: Isentropic transport within the Antarctic polar night vortex: Rossby wave breaking evidence and Lagrangian structures.
> - 10.1103/physreve.99.032221 may be a valid DOI for title: Unveiling the chaotic structure in phase space of molecular systems using Lagrangian descriptors
> - 10.1142/s0218127416300366 may be a valid DOI for title: Lagrangian Descriptors for Stochastic Differential Equations: A Tool for Revealing the Phase Portrait of Stochastic Dynamical Systems
> - 10.1039/c7cp05912d may be a valid DOI for title: Detecting reactive islands using Lagrangian descriptors and the relevance to transition path sampling
> - 10.1016/j.cplett.2017.09.008 may be a valid DOI for title: Obtaining time-dependent multi-dimensional dividing surfaces using Lagrangian descriptors
> - 10.1063/1.4997379 may be a valid DOI for title: Chemical dynamics between wells across a time-dependent barrier: Self-similarity in the Lagrangian descriptor and reactive basins
> - 10.1103/physrevlett.115.148301 may be a valid DOI for title: Lagrangian descriptors of thermalized transition states on time-varying energy surfaces
> - 10.1103/physreve.96.022222 may be a valid DOI for title: Lagrangian descriptors of driven chemical reaction manifolds
> - 10.1039/c5cp06624g may be a valid DOI for title: Deconstructing field-induced ketene isomerization through Lagrangian descriptors
> - 10.1007/s12043-008-0103-3 may be a valid DOI for title: Visual explorations of dynamics: The standard map
> - 10.1086/109234 may be a valid DOI for title: The Applicability of the Third Integral Of Motion: Some Numerical Experiments
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.ijmecsci.2017.10.040 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.cnsns.2019.105033 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.cplett.2020.137199 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.cnsns.2020.105331 is INVALID because of 'https://doi.org/' prefix
> ```

Fixed DOIs in `paper.bib`
