### 2021-07-23 #3482 @nicoguaro
I followed the installation described in the README and got the following

```bash
(ldds_joss) nguarinz@cocorna ~/ldds $ python -m unittest

Ran 0 tests in 0.000s

OK
```
