### 2021-07-26 #3482 @vkrajnak
> I followed the installation described in the README and got the following
>
> ```shell
> (ldds_joss) nguarinz@cocorna ~/ldds $ python -m unittest
>
> ----------------------------------------------------------------------
> Ran 0 tests in 0.000s
>
> OK
> ```

@nicoguaro Thank you for finding the unittest issue. In case it is acceptable to run `unittest` from a subfolder of the package, I've amended the README file to say

```bash
cd ldds/tests
python -m unittest
```

Then it works.
