### 2021-08-10 #3482 @nicoguaro
> @richardjgowers @nicoguaro, just wondering if we are required to make any additional changes at this stage to help with the review? Thank you.

@Shibabrat, I have moved forward in the review. This is reflected in the boxes checked above. Currently, I am reading the companion book showed in the documentation.
