### 2021-08-16 #3482 @broncio123
Hi @jmbr hope everything is OK. Just wondering if there has been any progress on the review checklist.
This hasn't been updated since 12 July (5 weeks ago).

As @Shibabrat mentioned above, if there's anything we can help out with, let us know. We're happy to do so.

@richardjgowers have you heard anything from @jmbr by any chance?
It doesn't seem that he has made any comment at all in the thread.
