### 2021-08-22 #3482 @jmbr
I've checked off all the items in the review checklist. Here are some comments:

* [SLOCCount](https://dwheeler.com/sloccount/) says that there are 742 lines of code (LOC) while the guidelines for JOSS say that submissions below 1000 LOC should be flagged.
```bash
python_count `git ls-files | grep -E '\.py$'`
21 docs/conf.py
9 ldds/__init__.py
223 ldds/base.py
66 ldds/base_discrete.py
33 ldds/discrete_maps.py
56 ldds/hamiltonians.py
44 ldds/testing_scripts.py
103 ldds/tools.py
79 ldds/vector_fields.py
29 setup.py
27 tests/separable_quadratic_hamiltonian.py
45 tests/test_uncoup_quad_ham2dof.py
7 tests/utils.py
Total:
742
```
* Sphinx gives some warnings while building the documentation.
```bash
$ make html
Running Sphinx v4.1.2
loading translations [python]... not available for built-in messages
WARNING: html_static_path entry '_static' does not exist
loading intersphinx inventory from https://docs.python.org/3/objects.inv...
building [mo]: targets for 0 po files that are out of date
building [html]: targets for 8 source files that are out of date
updating environment: [new config] 8 added, 0 changed, 0 removed
reading sources... [100%] user_guide
/workspace/math/ldds/docs/examples.rst:4: WARNING: Unknown directive type "examples".

.. examples::
/workspace/math/ldds/docs/examples.rst:20: WARNING: Explicit markup ends without a blank line; unexpected unindent.
/workspace/math/ldds/docs/examples.rst:96: WARNING: Inline interpreted text or phrase reference start-string without end-string.
/workspace/math/ldds/ldds/base.py:docstring of base.fit_vector_field:6: WARNING: Unexpected indentation.
/home/jmbr/python-venvs/ldds/lib/python3.8/site-packages/ldds-0.1.0-py3.8.egg/ldds/base.py:docstring of ldds.base.fit_vector_field:6: WARNING: Unexpected indentation.
looking for now-outdated files... none found
pickling environment... done
checking consistency... /workspace/math/ldds/docs/modules.rst: WARNING: document isn't included in any toctree
/workspace/math/ldds/docs/introduction.rst:59: WARNING: Citation [junginger2016lagrangian] is not referenced.
done
preparing documents... done
writing output... [100%] user_guide
/workspace/math/ldds/docs/introduction.rst:6: WARNING: citation not found: demian2017
generating indices... genindex py-modindex done
highlighting module code... [100%] vector_fields
writing additional pages... search done
copying static files... done
copying extra files... done
dumping search index in English (code: en)... done
dumping object inventory... done
build succeeded, 9 warnings.
```
* In the documentation, the function :math:`\mathcal{L}^{(\cdot)}(\mathbf{x}_0, t_0, \tau)` is used prior to being defined.
* There are typos in the documentation (enconded, accummulating, etc.). I recommend you spell check the docs and the Jupyter notebooks.
* I miss some structure in the code. For instance, I'd advice that you aim for an interface to solvers (e.g., Euler-Maruyama) that mirrors that of scipy.integrate or, perhaps better, rely on `sdeint` (https://github.com/mattja/sdeint/) as a dependency.
* Neither `fit_pes` nor `fit_vector_field` work after installation because the HDF5 files are not copied to the appropriate paths outside the source tree.
