### 2021-08-24 #3482 @broncio123
Thanks a lot for your feedback @jmbr
We will address each of your points promptly.

Hey @nicoguaro , any further progress on reviewing the functionality, documentation, and paper?
If there's anything we can help put with let us know.
