### 2021-08-24 #3482 @nicoguaro
I have marked most of the checklist above. Following are some comments that I would add to @jmbr's ones.

- Regarding the size (lines of code) of the package. I think that it is OK. The guideline suggests that number as a rule of thumb. In this case, I would say that it is clear the package has scientific value.

- For me, the documentation is confusing.

  - There isn't an example in the doc site, I was expecting one in the section "Getting started" or maybe on "Examples". I know that the examples are provided as notebooks, but having an introductory example in the docs helps a first-time user.
  - It is common to have the documentation for the API organized in modules. There is a file named ``modules.rst`` but it is not invoked in ``index.rst``.
  - The format required in ``vector_field`` in the function ``compute_lagrange_descriptor`` is not clear enough in the docstrings. Although, you get the idea in ``tutorial-4.ipynb``.
  - I tried to compute the descriptor for a "user-defined" system for a simple pendulum for a 50×50 grid in [-4, 4]² and ``tau=1`` but it ran for 5 hours without finishing. Maybe the selection of parameters is sub-optimal for this problem, but I couldn't find how to pick them in the documentation.
