+++ docs/examples.rst
@@ -1,7 +1,8 @@
+========
 Examples
 ========
 
-.. examples::
+Some examples used in the tutorial and considered as benchmark systems in dynamical systems.
 
 
 Discrete systems
@@ -17,6 +18,7 @@ kicked pendulum. Its equations of motion are given by the expressions:
     x_{n+1} &= x_{n} + y_{n} - \dfrac{K}{2\pi} \sin(2\pi x_{n)),
     y_{n+1} &= y_{n} - \dfrac{K}{2\pi} \sin(2\pi x_{n)),
    \end{align}
+
 where :math:`K` is the parameter that controls the forcing strength of the perturbation.
    
 The inverse map is described by:
