### 2021-08-27 5154bc3
Updated description of vector_field in compute_lagrangian_descriptor
