### 2021-08-28 #3482 @richardjgowers
Hi @nicoguaro @jmbr thanks for your work in reviewing things.  This seems like great feedback to improve the work.  Regarding the SLOC criteria, this is a guideline and the editors have decided that this package is suitable.
