### 2021-09-01 #3482 @broncio123
Hi @jmbr , I'm probably not understanding your comment

> * I miss some structure in the code. For instance, I'd advice that you aim for an interface to solvers (e.g., Euler-Maruyama) that mirrors that of scipy.integrate or, perhaps better, rely on `sdeint` (https://github.com/mattja/sdeint/) as a dependency.

By mirroring the interface, do you mean that for the function `base.EulerMaruyama_solver` the ordering and the format of the input parameters

 `(t_initial, u_initial, vector_field, ...)`

 should match the ordering and the format of the input parameters of  something like

 `scipy.integrate.solve_ivp(fun, t_span, y0, method='RK45', t_eval=None, dense_output=False, events=None, vectorized=False, args=None, **options)`?

Is this correct? otherwise, please, clarify.
