### 2021-09-06 #3482 @broncio123
> > Hi @nicoguaro , as for the last point, could you share with us a jupyter notebook or script with the setup of your simulation? Thanks! :)
>
> The notebook is in the following link: https://gist.github.com/nicoguaro/9dcd8900427e923a077bd0617dfa7bd2

Hi @nicoguaro

Thanks! I made some small tweaks to your setup here

https://github.com/champsproject/ldds/blob/test_pendulum/tutorials/demo/simple_pendulum.ipynb

Let us know if it worked.
b
