### 2021-09-08 #3482 @broncio123
Hi @jmbr , I hope you're doing well.

I wondering if you could help me with this query of mine, please.
https://github.com/openjournals/joss-reviews/issues/3482#issuecomment-910322939

If there's any other issue, please, just give us a shout and we'll look into it :)
