### 2021-09-08 #3482 @broncio123
> > Hi @nicoguaro
> > Thanks! I made some small tweaks to your setup here
> > https://github.com/champsproject/ldds/blob/test_pendulum/tutorials/demo/simple_pendulum.ipynb
>
> > Let us know if it worked.
>
> Yes, it worked. It seems that I missed some details but it works and for larger tau I get what is expected from that system.

Awesome! 😎
