+++ docs/getting_started.rst
@@ -1,15 +1,0 @@
-Getting started
-===============
-
-The purpose of this guide is to illustrate the computation and visualization of Lagrangian descriptors of dynamical systems. 
-
-To use this package, the prerequisite is a basic understanding of nonlinear dynamical systems concepts mentioned in the glossary.
-
-Then, a brief introduction to the method of Lagrangian descriptor can be found in the Tutorial section.
-
-Then, consider running the examples starting with the ones described in the Examples section.
-
-
-
-
-
