+++ docs/modules.rst
@@ -1,5 +1,6 @@
-LDDS
-====
+API Documentation
+=================
+
 
 .. toctree::
    :maxdepth: 4
