### 2021-09-10 #3482 @Shibabrat
Hi @nicoguaro, we've updated the documentation with a new structure and hope this is less confusing. Please have a look around at https://ldds.readthedocs.io/en/latest/ and let us know if any other formatting change/s that will improve readability. Thanks.
