### 2021-09-10 #3482 @nicoguaro
> Hi @nicoguaro, we've updated the documentation with a new structure and hope this is less confusing. Please have a look around at https://ldds.readthedocs.io/en/latest/ and let us know if any other formatting change/s that will improve readability. Thanks.

I think that this new structure works better. I see some placeholders but I guess that you will fill them while completing the reviewing.

@richardjgowers, I have marked all the boxes above and consider that we can move forward with publication.
