### 2021-09-21 #3482 @richardjgowers
@broncio123 @Shibabrat paper looks good now.  Can you update the software repo to match the v0.1.0 version that this paper references?  After that if you can create a release and archive that on something like zenodo and then report the DOI of that archive we can proceed.
