### 2021-09-21 #3482 @broncio123
> @broncio123 @Shibabrat paper looks good now. Can you update the software repo to match the v0.1.0 version that this paper references? After that if you can create a release and archive that on something like zenodo and then report the DOI of that archive we can proceed.

The software is up-to-date and now a release is available here
https://github.com/champsproject/ldds/releases/tag/v0.1.0

I have also archived the source code of this release on ZENODO with URL
https://zenodo.org/record/5519580

Let me know if there's anything else we need to do.

b
