### 2021-09-22 #3482 @richardjgowers
@broncio123 sorry I should have been clearer, I think you've manually set the DOI on the zenodo archive to the JOSS DOI? I need the zenodo archive to have its own DOI (zenodo will auto generate one iirc) then the paper will point to that DOI (associated with the software archive) to generate the DOI associated with the paper.
