### 2021-09-23 #3482 @broncio123
> @broncio123 sorry I should have been clearer, I think you've manually set the DOI on the zenodo archive to the JOSS DOI? I need the zenodo archive to have its own DOI (zenodo will auto generate one iirc) then the paper will point to that DOI (associated with the software archive) to generate the DOI associated with the paper.

No problem, @richardjgowers
The ZENODO DOI is:
> 10.5281/zenodo.5519580

The JOSS DOI no longer appears on the Zenodo record https://zenodo.org/record/5519580
