### 2021-09-23 #3482 @kthyng
Hi @Shibabrat! Just looked through your paper. It looks good except for the references. Throughout the paper, you have used `[@reference], [@reference2]` for inline lists of references instead of `@reference, @reference2`, etc. Please fix all of the times this happens in the paper. Also I see some capitalization mistakes in the references list. For example, Craven and Hernandez doesn't have "Lagrangian" capitalized. You can preserve capitalization in your .bib file with {} around strings.
