+++ paper/paper.bib
@@ -147,7 +147,7 @@
 
 @book{ldbook2020,
     author = {Agaoglou, M. and Aguilar-Sanjuan, B. and Garc{\'i}a-Garrido, V. J. and Gonz{\'a}lez-Montoya, F. and Katsanikas, M. and  Kraj\v{n}{\'a}k, V. and Naik, S. and Wiggins, S.},
-    title = {Lagrangian Descriptors: Discovery and Quantification of Phase Space Structure and Transport},
+    title = {{Lagrangian Descriptors: Discovery and Quantification of Phase Space Structure and Transport}},
     year = {2020},
     doi = {10.5281/zenodo.3958985},
     publisher = {zenodo: 10.5281/zenodo.3958985},
@@ -393,7 +393,7 @@
 	Journal = {Regular and Chaotic Dynamics},
 	Number = {5},
 	Pages = {551--568},
-	Title = {The Application of Lagrangian Descriptors to 3D Vector Fields},
+	Title = {{The Application of Lagrangian Descriptors to 3D Vector Fields}},
 	Volume = {23},
 	Year = {2018},
 	Bdsk-Url-1 = {https://doi.org/10.1134/S1560354718050052}}
@@ -414,7 +414,7 @@
 	Number = {3},
 	Pages = {032221},
 	Publisher = {APS},
-	Title = {Unveiling the chaotic structure in phase space of molecular systems using Lagrangian descriptors},
+	Title = {{Unveiling the chaotic structure in phase space of molecular systems using Lagrangian descriptors}},
 	Volume = {99},
 	Year = {2019}}
 	
@@ -427,7 +427,7 @@
 	Number = {13},
 	Pages = {1630036},
 	Publisher = {World Scientific},
-	Title = {Lagrangian Descriptors for Stochastic Differential Equations: A Tool for Revealing the Phase Portrait of Stochastic Dynamical Systems},
+	Title = {{Lagrangian Descriptors for Stochastic Differential Equations: A Tool for Revealing the Phase Portrait of Stochastic Dynamical Systems}},
 	Volume = {26},
 	Year = {2016}}
 	
@@ -440,7 +440,7 @@
 	Number = {7},
 	Pages = {4970--4981},
 	Publisher = {Royal Society of Chemistry},
-	Title = {Detecting reactive islands using Lagrangian descriptors and the relevance to transition path sampling},
+	Title = {{Detecting reactive islands using Lagrangian descriptors and the relevance to transition path sampling}},
 	Volume = {20},
 	Year = {2018}}
 	
@@ -450,7 +450,7 @@
     doi = {10.1016/j.cplett.2017.09.008},
 	Pages = {194--199},
 	Publisher = {Elsevier},
-	Title = {Obtaining time-dependent multi-dimensional dividing surfaces using Lagrangian descriptors},
+	Title = {{Obtaining time-dependent multi-dimensional dividing surfaces using Lagrangian descriptors}},
 	Volume = {687},
 	Year = {2017}}
 
@@ -461,7 +461,7 @@
 	Number = {6},
 	Pages = {064101},
 	Publisher = {AIP Publishing},
-	Title = {Chemical dynamics between wells across a time-dependent barrier: Self-similarity in the Lagrangian descriptor and reactive basins},
+	Title = {{Chemical dynamics between wells across a time-dependent barrier: Self-similarity in the Lagrangian descriptor and reactive basins}},
 	Volume = {147},
 	Year = {2017}}
 	
@@ -472,7 +472,7 @@
 	Number = {14},
 	Pages = {148301},
 	Publisher = {APS},
-	Title = {Lagrangian descriptors of thermalized transition states on time-varying energy surfaces},
[…]
