### 2021-09-23 #3482 @Shibabrat
@kthyng thank you for the feedback. I've changed the typesetting for references throughout and used `()` for enclosing references to improve readability.
