+++ paper/paper.bib
@@ -26,7 +26,7 @@
 
 
 @article{mackay_transport_1984,
-	title={Transport in Hamiltonian systems},
+	title={Transport in {H}amiltonian systems},
 	author={MacKay, R. S. and Meiss, J. D. and Percival, I. C.},
 	journal={Physica D: Nonlinear Phenomena},
 	volume={13},
@@ -271,7 +271,7 @@
 	Issn = {1007-5704},
 	Journal = {Communications in Nonlinear Science and Numerical Simulation},
 	Pages = {105331},
-	Title = {Exploring isomerization dynamics on a potential energy surface with an index-2 saddle using lagrangian descriptors},
+	Title = {{Exploring isomerization dynamics on a potential energy surface with an index-2 saddle using lagrangian descriptors}},
 	Volume = {89},
 	Year = {2020}
     }
@@ -279,7 +279,7 @@
 @article{krajnak2020manifld,
 	Author = {Kraj{\v{n}}{\'a}k, V. and Ezra, G. S. and Wiggins, S.},
 	Journal = {International Journal of Bifurcation and Chaos},
-	Title = {Using Lagrangian Descriptors to Uncover Invariant Structures in Chesnavich’s Isokinetic Model with Application to Roaming},
+	Title = {{Using Lagrangian Descriptors to Uncover Invariant Structures in Chesnavich’s Isokinetic Model with Application to Roaming}},
 	Year = {2020},
 	volume={30},
 	number={5},
@@ -294,7 +294,7 @@
 	Number = {23},
 	Pages = {235702},
 	Publisher = {{IOP} Publishing},
-	Title = {Revealing roaming on the double Morse potential energy surface with Lagrangian descriptors},
+	Title = {Revealing roaming on the double {M}orse potential energy surface with Lagrangian descriptors},
 	Url = {https://doi.org/10.1088/1751-8121/ab8b75},
 	Volume = {53},
 	Year = 2020,
@@ -339,7 +339,7 @@
 	Issue = {2},
 	Journal = {Phys. Rev. E},
 	Pages = {022204},
-	Title = {Finding normally hyperbolic invariant manifolds in two and three degrees of freedom with {H}{\'e}non-{H}eiles type potential},
+	Title = {Finding normally hyperbolic invariant manifolds in two and three degrees of freedom with {H}{\'e}non-{H}eiles-type potential},
 	Volume = {100},
 	Year = {2019},
 	Bdsk-Url-1 = {https://doi.org/10.1103/PhysRevE.100.022204}}
@@ -373,7 +373,7 @@
 	Journal = {J. Atmos. Sci.},
     doi = {10.1175/JAS-D-12-0274.1},
 	Pages = {2982-3001},
-	Title = {{Isentropic transport within the Antarctic polar night vortex: Rossby wave breaking evidence and Lagrangian structures.}},
+	Title = {{Isentropic Transport within the Antarctic Polar-Night Vortex: Rossby Wave Breaking Evidence and Lagrangian Structures}},
 	Volume = {70},
 	Year = {2013}}
 
@@ -536,7 +536,7 @@
     author = {Coulliette, C. and Wiggins, S.},
     journal = {Nonlinear Processes in Geophysics},
     pages = {69--94},
-    title = {{Nonlinear Processes in Geophysics Intergyre transport in a wind-driven, quasigeostrophic double gyre: An application of lobe dynamics}},
+    title = {Nonlinear Processes in Geophysics Intergyre transport in a wind-driven, quasigeostrophic double gyre: An application of lobe dynamics},
     volume = {8},
     year = {2001}
 }
@@ -600,7 +600,7 @@
[…]
