+++ paper/paper.bib
@@ -184,7 +184,7 @@
 }
 
 @article{libcfd2lcs,
-	title = {Integrated computation of finite time {L}yapunov exponent fields during direct numerical simulation of unsteady flows},
+	title = {Integrated Computation of Finite Time Lyapunov Exponent Fields During Direct Numerical Simulation of Unsteady Flows.},
 	author = {Finn, J. and Apte, S. V.},
 	journal = {Chaos},
 	volume = {23},
@@ -195,7 +195,7 @@
 }
 
 @misc{lcsmatlabkit,
-	title = {{LCS M}atlab kit},
+	title = {{LCS MATLAB} kit},
 	version = {2.3},
 	author = {{Dabiri Lab}},
 	year = {2009},
@@ -211,7 +211,7 @@
 }
 
 @article{activeBarriers,
-	title = {Objective barriers to the transport of dynamically active vector fields},
+	title = {{Objective Barriers to the Transport of Dynamically Active Vector Fields.}},
 	author = {Haller, G. and Katsanoulis, S. and Holzner, M. and Frohnapfel, B. and Gatti, D.},
 	year = {2020},
 	journal = {J. Fluid Mech.},
@@ -260,7 +260,7 @@
 	Journal = {International Journal of Bifurcation and Chaos},
 	Number = {04},
 	Pages = {2030008},
-	Title = {Tilting and Squeezing: {P}hase space geometry of {H}amiltonian saddle-node bifurcation and its influence on chemical reaction dynamics},
+	Title = {{Tilting and Squeezing: Phase Space Geometry of Hamiltonian Saddle-Node Bifurcation and Its Influence on Chemical Reaction Dynamics.}},
 	Volume = {30},
 	Year = {2020},
 	Bdsk-Url-1 = {https://doi.org/10.1142/S0218127420300086}}
@@ -271,7 +271,7 @@
 	Issn = {1007-5704},
 	Journal = {Communications in Nonlinear Science and Numerical Simulation},
 	Pages = {105331},
-	Title = {{Exploring isomerization dynamics on a potential energy surface with an index-2 saddle using lagrangian descriptors}},
+	Title = {{Exploring Isomerization Dynamics on a Potential Energy Surface with an Index-2 Saddle Using Lagrangian Descriptors.}},
 	Volume = {89},
 	Year = {2020}
     }
@@ -316,7 +316,7 @@
 	Journal = {Climate Dynamics},
 	Number = {5},
 	Pages = {2779--2792},
-	Title = {Lagrangian study of the final warming in the southern stratosphere during 2002: Part I. The vortex splitting at upper levels},
+	Title = {{Lagrangian Study of the Final Warming in the Southern Stratosphere During 2002: Part I. The Vortex Splitting at Upper Levels.}},
 	Url = {https://doi.org/10.1007/s00382-019-04832-y},
 	Volume = {53},
 	Year = {2019},
@@ -328,7 +328,7 @@
 	Journal = {Climate Dynamics},
 	Number = {3},
 	Pages = {1277--1286},
-	Title = {Lagrangian study of the final warming in the southern stratosphere during 2002: Part II. 3D structure},
+	Title = {{Lagrangian Study of the Final Warming in the Southern Stratosphere During 2002: Part II. 3D Structure.}},
 	Url = {https://doi.org/10.1007/s00382-019-04833-x},
 	Volume = {53},
 	Year = {2019}}
@@ -352,7 +352,7 @@
 	Journal = {International Journal of Bifurcation and Chaos},
 	Number = {14},
 	Pages = {1750225},
[…]
