+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{meiss_symplectic_1992,
-	title = {Symplectic maps, variational principles, and transport},
+	title = {{Symplectic Maps, Variational Principles, and Transport.}},
 	author = {Meiss, J. D.},
 	journal = {Rev. Mod. Phys.},
 	volume = {64},
@@ -13,7 +13,7 @@
 
 
 @article{waalkens_escape_2005,
-	title = {Escape from planetary neighbourhoods},
+	title = {{Escape From Planetary Neighbourhoods.}},
 	volume = {361},
 	doi = {10.1111/j.1365-2966.2005.09237.x},
 	number = {3},
@@ -26,7 +26,7 @@
 
 
 @article{mackay_transport_1984,
-	title={Transport in {H}amiltonian systems},
+	title={{Transport in Hamiltonian Systems.}},
 	author={MacKay, R. S. and Meiss, J. D. and Percival, I. C.},
 	journal={Physica D: Nonlinear Phenomena},
 	volume={13},
@@ -39,7 +39,7 @@
 
 
 @article{koon_heteroclinic_2000,
-	title = {Heteroclinic connections between periodic orbits and resonance transitions in celestial mechanics},
+	title = {{Heteroclinic Connections between Periodic Orbits and Resonance Transitions in Celestial Mechanics.}},
 	volume = {10},
 	doi = {10.1063/1.166509},
 	number = {2},
@@ -51,7 +51,7 @@
 }
 
 @article{zhong_tube_2018,
-	title = {A tube dynamics perspective governing stability transitions: An example based on snap-through buckling},
+	title = {{A Tube Dynamics Perspective Governing Stability Transitions: An Example Based on Snap-Through Buckling.}},
 	journal = {International Journal of Mechanical Sciences},
 	volume = {149},
 	pages = {413-428},
@@ -61,7 +61,7 @@
 }
 
 @article{zhong_geometry_2020,
-	title = {Geometry of escape and transition dynamics in the presence of dissipative and gyroscopic forces in two degree of freedom systems},
+	title = {{Geometry of Escape and Transition Dynamics in the Presence of Dissipative and Gyroscopic Forces in Two Degree of Freedom Systems.}},
 	journal = {Communications in Nonlinear Science and Numerical Simulation},
 	volume = {82},
 	pages = {105033},
@@ -72,7 +72,7 @@
 
 @Article{meiss15,
   author	= {Meiss, J. D.},
-  title		= {Thirty years of turnstiles and transport},
+  title		= {{Thirty Years of Turnstiles and Transport.}},
   journal	= {Chaos},
   volume	= {25},
   number	= {9},
@@ -82,7 +82,7 @@
 
 @software{jupyterbook_2020,
 	author       = {{Executable Books Community}},
-	title        = {Jupyter Book},
+	title        = {{Jupyter Book.}},
 	month        = feb,
 	year         = 2020,
 	publisher    = {Zenodo},
@@ -92,7 +92,7 @@
 }
 
 @book{wiggins1994normally,
-	title= {Normally hyperbolic invariant manifolds in dynamical systems},
+	title= {{Normally Hyperbolic Invariant Manifolds in Dynamical Systems.}},
 	author= {Wiggins, S.},
 	volume= {105},
 	year= {1994},
@@ -103,8 +103,7 @@
 @article{OzoriodeAlmeida1990,
 	author = "Ozorio de Almeida, A. M.  and De Leon, N.  and Mehta, M. A.  
 	and Marston, C. C.",
-	title = "Geometry and dynamics of stable and unstable cylinders in 
-	Hamiltonian systems",
[…]
