+++ paper/paper.bib
@@ -183,7 +183,7 @@
 }
 
 @article{libcfd2lcs,
-	title = {Integrated Computation of Finite Time Lyapunov Exponent Fields During Direct Numerical Simulation of Unsteady Flows.},
+	title = {{Integrated Computation of Finite Time Lyapunov Exponent Fields During Direct Numerical Simulation of Unsteady Flows.}},
 	author = {Finn, J. and Apte, S. V.},
 	journal = {Chaos},
 	volume = {23},
@@ -554,7 +554,7 @@
 @article{Chirikov1969,
 	author = {Chirikov, B. V.},
 	publisher = {CERN},
-	title = {Research Concerning the Theory of Non-Linear Resonance and Stochasticity [translated at {CERN by A. T. S}anders]},
+	title = {{Research Concerning the Theory of Non-Linear Resonance and Stochasticity [translated at CERN by A. T. Sanders]}},
 	journal = {{CERN} trans.},
 	volume = {71},
 	number = {40},
