+++ paper/paper.bib
@@ -498,7 +498,7 @@
 
 @article{Henon1964,
     author = {Henon, M. and Heiles, C.},
-    journal = {THE ASTRONOMICAL JOURNAL},
+    journal = {The Astronomical Journal},
     doi = {10.1086/109234},
     number = {1},
     title = {{The Applicability of the Third Integral of Motion: Some Numerical Experiments.}},
