### 2021-09-28 #3482 @broncio123
> 🚨🚨🚨 **THIS IS NOT A DRILL, YOU HAVE JUST ACCEPTED A PAPER INTO JOSS!** 🚨🚨🚨
>
> Here's what you must now do:
>
> 1. Check final PDF and Crossref metadata that was deposited 👉 [Creating pull request for 10.21105.joss.03482 joss-papers#2626](https://github.com/openjournals/joss-papers/pull/2626)
> 2. Wait a couple of minutes, then verify that the paper DOI resolves https://doi.org/10.21105/joss.03482
> 3. If everything looks good, then close this review issue.
> 4. Party like you just published a paper! 🎉🌈🦄💃👻🤘
>
> Any issues? Notify your editorial technical team...

Thank you very much everybody for your time, dedication, and feedback during the review process!
@richardjgowers @jmbr @nicoguaro @danielskatz @kthyng

The authors.
